Put together in a single file related small classes
