--- src/Symtab/Value/BoolValue.h
@@ -1,16 +1,0 @@
-//
-// Created by Roger Generoso Masós on 29/03/2020.
-//
-
-#ifndef CSP2SAT_BOOLVALUE_H
-#define CSP2SAT_BOOLVALUE_H
-
-#include "Value.h"
-
-namespace GOS {
-
-
-
-}
-
-#endif //CSP2SAT_BOOLVALUE_H
