--- src/Symtab/Value/IntValue.h
@@ -1,15 +1,0 @@
-//
-// Created by Roger Generoso Masós on 29/03/2020.
-//
-
-#ifndef CSP2SAT_INTVALUE_H
-#define CSP2SAT_INTVALUE_H
-
-#include "Value.h"
-
-namespace GOS {
-
-
-}
-
-#endif //CSP2SAT_INTVALUE_H
